requirements.txt
streamlit==1.29.0
numpy==1.26.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
umap-learn==0.5.3
plotly==5.18.0
python-dotenv==1.0.0